+++ server/router.ts
@@ -25,9 +25,10 @@ import { monitorRoutes } from "./routes/monitor";
 import { supportRoutes } from "./routes/support";
 import { paymentRoutes } from "./routes/payment";
 import { opsRoutes } from "./routes/ops";
+import { telemetryRoutes } from "./routes/telemetry";
 import { pageRoutes } from "./routes/pages";
 
-const ROUTES: RouteFn[] = [metaRoutes, geoRoutes, assessRoutes, analysisRoutes, cropsRoutes, regionFitRoutes, journalRoutes, accountRoutes, marketRoutes, budgetRoutes, guideRoutes, foreignRoutes, notifyRoutes, pushRoutes, alertsRoutes, monitorRoutes, supportRoutes, paymentRoutes, opsRoutes, pageRoutes];
+const ROUTES: RouteFn[] = [metaRoutes, geoRoutes, assessRoutes, analysisRoutes, cropsRoutes, regionFitRoutes, journalRoutes, accountRoutes, marketRoutes, budgetRoutes, guideRoutes, foreignRoutes, notifyRoutes, pushRoutes, alertsRoutes, monitorRoutes, supportRoutes, paymentRoutes, opsRoutes, telemetryRoutes, pageRoutes];
 
 export async function route(ctx: Ctx, req: http.IncomingMessage, res: http.ServerResponse, url: URL): Promise<void> {
   for (const handler of ROUTES) {